Redo styles
It was a nice screenshot, but the header was ugly as sin on the page
per se. And it didn't match the main site, or feel at all one with it.
Which, I mean, it's too small an app to have its own identity like that.
In with the new.

* Add diagonal grid background
* Use Red Sox colorscheme
* Move board to right and all buttons to left
  - Board come before buttons in DOM with float: right; so that when
    things get squishy (900px) it goes above buttons.
* Style buttons
* Redo select tag
  - Give puzzle number in <option>s, not string
  - Use <optgroup label>s to display puzzles by row with separators
  - Put click handler on <optgroup> tags for more intuitive selecting